+++ lib/questionnaires/model/src/questionnaire_model_defaults.dart
@@ -22,6 +22,11 @@ class QuestionnaireModelDefaults {
   /// Returns whether questions should offer an option to not answer the question
   final bool implicitNullOption;
 
+  /// Returns whether boolean items should be tri-state.
+  ///
+  /// Boolean items will be tri-state if [true], or bi-state if [false].
+  final bool booleanTriState;
+
   /// Returns a prefix, starting with 1, that provides the number
   /// of the [QuestionItemModel] within the ordered sequence of [QuestionItemModels].
   ///
@@ -56,5 +61,6 @@ class QuestionnaireModelDefaults {
     this.prefixBuilder,
     this.disabledDisplay = QuestionnaireDisabledDisplay.hidden,
     this.implicitNullOption = true,
+    this.booleanTriState = false,
   });
 }